+++ src/application/components/ObtenerValorInicial.jsx
