+++ src/application/components/styled.js
@@ -1,0 +1,67 @@
+import styled from "styled-components";
+
+export const StyleText = styled.label `
+    font-family: Arial, Helvetica, sans-serif;
+`
+export const StyleP = styled.p `
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight:bold;
+    font-size: 20px;
+    color: coral;
+`
+export const StyleH1 = styled.h1 `
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight:bold;
+    font-size: 40px;
+`
+export const StyleServiciosWeb = styled.div `
+    border: 3px solid black;
+    border-radius: 15px;
+    width: 270px;
+    margin: 0 auto;
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+    font-family: Arial, Helvetica, sans-serif;
+`;
+
+export const StyleSDiv = styled.div `
+    background-color: coral;
+    border-radius: 15px;
+    width: fit-content;
+    margin: 0 auto;
+    margin-top: 1rem;
+    padding: 1rem;
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight:bold;
+`;
+
+export const InputStyle = styled.input `
+    width: 35px;
+    border: none;
+    text-align:center;
+`;
+
+export const ButtonsStyle = styled.button `
+    background-color: coral;
+    border: none;
+    border-radius: 5px;
+    color: white;
+    font-size: 20px;
+    font-weight:bold;
+    width: 25px;
+`
+export const HomeStyle = styled.div `
+    margin: 0 auto;
+    text-align:center;
+    font-size: 20px;
+    
+`
+
+export const ButtonHome = styled.button `
+    background-color: coral;
+    border: none;
+    border-radius: 5px;
+    padding: 10px;
+    font-size: 20px;
+    font-weight:bold;
+`
