--- src/components/styled.js
@@ -1,26 +1,0 @@
-import styled from "styled-components";
-
-export const StyleServiciosWeb = styled.div `
-    border: 3px solid black;
-    border-radius: 15px;
-    width: 270px;
-    margin: 0 auto;
-    margin-top: 1rem;
-    margin-bottom: 1rem;
-`;
-
-export const InputStyle = styled.input `
-    width: 35px;
-    border: none;
-    text-align:center;
-`;
-
-export const ButtonsStyle = styled.button `
-    background-color: coral;
-    border: none;
-    border-radius: 5px;
-    color: white;
-    font-size: 20px;
-    font-weight:bold;
-    width: 25px;
-`
